backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
playwright==1.47.0
google-genai==0.8.0
pillow==10.4.0
python-multipart==0.0.9
pydantic==2.8.2
python-dotenv==1.0.1
httpx==0.27.2
